requirements.txt
loguru==0.7.3
paho-mqtt==2.1.0
opencv-python==4.8.1.78
numpy==1.26.2